fixed a bug in uniform_int_random where I asked for <double>

+++ src/Utilities.cpp
@@ -133,7 +133,7 @@ Distribution make_distribution(json const &j) {
 if(distributionName == #name_) return make_distribution_impl<std::name_<type_>>(j, ## __VA_ARGS__);
 
     // Continious Distributions
-    generate_distribution_factory(uniform_int_distribution, double, "a", "b");
+    generate_distribution_factory(uniform_int_distribution, int, "a", "b");
     generate_distribution_factory(uniform_real_distribution, double, "a", "b");
     generate_distribution_factory(binomial_distribution, double, "p", "t");
     generate_distribution_factory(exponential_distribution, double, "lambda");
